refactor(server): extract plan-run coordinator wiring into sibling module
Decompose src/server/main/index.ts (was 547 lines) below the 500-line
global limit by extracting the plan-run coordinator wiring block into a
new sibling src/server/main/plan-run-wiring.ts, mirroring the existing
*-wiring.ts precedent (preview-/detector-/observability-wiring). Adds
bootPlanRunCoordinatorWiring() which loads config, constructs the
coordinator (incl. reopenPr + transitionPlot closures), emits the
disabled/running log, and returns the handle for teardown.

index.ts drops to 440 lines; plan-run-wiring.ts is 213. The supervisor
entry path src/server/main/index.ts is unchanged (still exports
bootServer + import.meta.main). reopenPr refactored into helpers to stay
under the cognitive-complexity ceiling in the new (non-grandfathered) file.

warren-8e91
